refactor: separate styles, lazy load footer
