--- site/src/layouts/page.tsx
@@ -1,27 +1,0 @@
-import type { LayoutProps } from '@typings/layout_props'
-
-import { styled } from '@junat/design'
-
-import Footer from '@components/common/footer'
-
-const StyledPage = styled('div', {
-  paddingTop: '1.875rem',
-  width: '100%',
-  margin: 'auto'
-})
-
-const StyledContent = styled('div', {
-  paddingInline: '1.875rem',
-  margin: 'auto',
-  maxWidth: '500px',
-  minHeight: '100vh'
-})
-
-export default function Page({ children }: LayoutProps) {
-  return (
-    <StyledPage>
-      <StyledContent>{children}</StyledContent>
-      <Footer />
-    </StyledPage>
-  )
-}
