+++ site/src/layouts/page/styles.tsx
@@ -1,0 +1,14 @@
+import { styled } from '@junat/design'
+
+export const StyledPage = styled('div', {
+  paddingTop: '1.875rem',
+  width: '100%',
+  margin: 'auto'
+})
+
+export const StyledContent = styled('div', {
+  paddingInline: '1.875rem',
+  margin: 'auto',
+  maxWidth: '500px',
+  minHeight: '100vh'
+})
